this commit introduces several important changes, and is meant for investigation rather than intended for direct merging.
1. use the private server directory

This aleviates failures in multi-node, *even* when both the public and private
directories are at the same commit. I suspect this may be a packaging problem,
but have no direct proof of this.

2. test all (by default) or run a specific profile - convenience
3. consistent logging to the systest [operator] appender - convenience
4. fix non-matching clause in ..._utils:cluster/1
5. warnings as errors during compile (avoid stuff like 4)
6. demonstrate what appears to be a genuine regression in ha behaviour

in 6, it appears that the restarted_master_honours_declarations test case
demonstrates how, even without breaking a cluster, a queue redeclaration
can succeed that shouldn't. Perhaps this is not as broken as it appears to
be, but I cannot see any merge between the HA policy branch (24908) and
the direct ancestor of this one (24914) or default, so I'm assuming it
is a bug of sorts.