make prepare: Use the testsuite environment to determine remote node name
By using the testsuite settings and asking Erlang about the computed
nodename, we make sure `make prepare` will use the correct `rabbit@...`
while trying to create the cluster.

It helps on hosts such as Travis CI's ones where a `$NODENAME` is
forced in /etc/rabbitmq/rabbitmq-env.conf. The testsuite ignores this
`$NODENAME` because it explicitely sets one. Now, `make prepare` does
the same. They are now on the same page and the cluster is created.