maybe_delay/1: Abort the wait if the queue receives an EXIT signal
If a test fails and the node is being taken down, this infinite wait
blocks the termination of the node. Therefore, the testsuite never
returns.

Now, maybe_delay/1 aborts if there is a pending EXIT signal. It consumes
and requeues the signal so rabbit_amqqueue_process can handle it
normally and let the node terminates.