rabbit_config_schema_test: Do not depend on a particular file hierarchy
There are enough environment variables to query the location of various
bits. This fixes the test when the Umbrella is not in use.

There is still a problem: the test writes to the source directory
instead of a temporary location, preventing parallel execution and
leaving temporary data in the directory in case of a test failure.